conversational_bot/requirements.txt
langchain==0.2.1
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.35.0
